abstract out updateJob
