+++ src/jobOutput/cancel.js
@@ -1,6 +1,8 @@
+import { updateJob } from './updateJob';
+
 import { postCancelJobRequest } from '../h2oProxy/postCancelJobRequest';
 
-export function cancel(_, _key, _job, updateJob) {
+export function cancel(_, _key, _job) {
   return postCancelJobRequest(_, _key, (error, result) => {
     if (error) {
       return console.debug(error);
