+++ src/jobOutput/h2oJobOutput.js
@@ -1,12 +1,10 @@
-import { getJobOutputStatusColor } from './getJobOutputStatusColor';
-import { getJobProgressPercent } from './getJobProgressPercent';
+
 import { isJobRunning } from './isJobRunning';
 import { canView } from './canView';
 import { cancel } from './cancel';
+import { updateJob } from './updateJob';
 
 import { getJobRequest } from '../h2oProxy/getJobRequest';
-import { formatMilliseconds } from '../utils/formatMilliseconds';
-
 
 import { flowPreludeFunction } from '../flowPreludeFunction';
 const flowPrelude = flowPreludeFunction();
@@ -53,44 +51,6 @@ export function h2oJobOutput(_, _go, _job) {
     WARN: 'fa-warning orange',
     INFO: 'fa-info-circle',
   };
-  const updateJob = job => {
-    let cause;
-    let message;
-    let messages;
-    _runTime(formatMilliseconds(job.msec));
-    _progress(getJobProgressPercent(job.progress));
-    _remainingTime(job.progress ? formatMilliseconds(Math.round((1 - job.progress) * job.msec / job.progress)) : 'Estimating...');
-    _progressMessage(job.progress_msg);
-    _status(job.status);
-    _statusColor(getJobOutputStatusColor(job.status));
-    if (job.error_count) {
-      messages = (() => {
-        let _i;
-        let _len;
-        const _ref = job.messages;
-        const _results = [];
-        for (_i = 0, _len = _ref.length; _i < _len; _i++) {
-          message = _ref[_i];
-          if (message.message_type !== 'HIDE') {
-            _results.push({
-              icon: messageIcons[message.message_type],
-              message: `${message.field_name}: ${message.message}`,
-            });
-          }
-        }
-        return _results;
-      })();
-      _messages(messages);
-    } else if (job.exception) {
-      cause = new Error(job.exception);
-      if (job.stacktrace) {
-        cause.stack = job.stacktrace;
-      }
-      _exception(Flow.failure(_, new Flow.Error('Job failure.', cause)));
-    }
-    _canView(canView(_destinationType, job));
-    return _canCancel(isJobRunning(job));
-  };
   const refresh = () => {
     _isBusy(true);
     return getJobRequest(_, _key, (error, job) => {
@@ -99,7 +59,23 @@ export function h2oJobOutput(_, _go, _job) {
         _exception(Flow.failure(_, new Flow.Error('Error fetching jobs', error)));
         return _isLive(false);
       }
-      updateJob(job);
+      updateJob(
+        _,
+        _runTime,
+        _progress,
+        _remainingTime,
+        _progressMessage,
+        _status,
+        _statusColor,
+        messageIcons,
+        _messages,
+        _exception,
+        _canView,
+        canView,
+        _destinationType,
+        _canCancel,
+        job
+      );
       if (isJobRunning(job)) {
         if (_isLive()) {
           return lodash.delay(refresh, 1000);
@@ -140,7 +116,23 @@ export function h2oJobOutput(_, _go, _job) {
     }
   };
   const initialize = job => {
-    updateJob(job);
+    updateJob(
+      _,
+      _runTime,
+      _progress,
+      _remainingTime,
+      _progressMessage,
+      _status,
+      _statusColor,
+      messageIcons,
+      _messages,
+      _exception,
+      _canView,
+      canView,
+      _destinationType,
+      _canCancel,
+      job
+    );
     if (isJobRunning(job)) {
       return _isLive(true);
     }
@@ -165,7 +157,7 @@ export function h2oJobOutput(_, _go, _job) {
     isLive: _isLive,
     canView: _canView.bind(this, _destinationType),
     canCancel: _canCancel,
-    cancel: cancel.bind(this, _, _key, _job, updateJob),
+    cancel: cancel.bind(this, _, _key, _job),
     view,
     template: 'flow-job-output',
   };
